requirements.txt
aioodbc==0.5.0
alembic==1.11.1
altair==5.1.1
amqp==5.2.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astroid==2.15.6
asttokens==2.4.0
async-timeout==4.0.3
asyncpg==0.29.0
atomicwrites==1.4.1
attrs==23.1.0
autoflake==1.4
azure-core==1.29.3
azure-identity==1.14.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.5
bandit==1.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.3.0
blinker==1.6.2
blis==0.7.11
build==1.0.3
cachetools==5.3.1
catalogue==2.0.10
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.11
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.3
configs==3.0.3
coverage==7.4.0
cryptography==41.0.1
cssselect==1.2.0
cymem==2.0.8
databases==0.4.2
databricks-sql-connector==2.0.2
DBUtils==3.0.3
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docker-pycreds==0.4.0
docx2txt==0.8
duckdb==0.8.1
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
filelock==3.12.3
filetype==1.2.0
flake8==5.0.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
griffe==0.38.1
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kombu==5.3.4
langchain-experimental==0.0.14
langcodes==3.3.0
langsmith==0.0.33
lazy-object-proxy==1.9.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mike==1.1.2
mkdocs==1.5.2
mkdocs-autorefs==0.4.1
mkdocs-gen-files==0.5.0
mkdocs-glightbox==0.3.4
mkdocs-literate-nav==0.6.0
mkdocs-material==9.1.15
mkdocs-material-extensions==1.1.1
mkdocs-section-index==0.3.5
mkdocs-video==1.5.0
mkdocstrings==0.22.0
mkdocstrings-python==1.6.0
monotonic==1.6
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
msg-parser==1.2.0
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.3
olefile==0.46
onnxruntime==1.15.1
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
paginate==0.5.6
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pbr==5.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
pip==23.3.1
platformdirs==3.10.0
pluggy==1.3.0
portalocker==2.7.0
posthog==3.0.2
pre-commit-hooks==4.4.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.0
psycopg2==2.9.6
pulsar-client==3.3.0
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.20
pydantic==1.10.12
pydantic_core==2.6.3
pydantic-settings==2.0.3
pydeck==0.8.1b0
pyflakes==2.5.0
Pygments==2.16.1
pylint==2.17.5
pymdown-extensions==10.5
Pympler==1.0.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pypandoc==1.11
pyparsing==3.1.1
pypdf==3.9.1
pyphen==0.14.0
pyproject_hooks==1.0.0
pyquery==2.0.0
pyreadline3==3.4.1
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==23.2.1
readtime==3.0.0
redis==5.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.0.284
scikit-learn==1.2.2
scipy==1.11.2
sentry-sdk==1.36.0
setproctitle==1.3.3
setuptools==68.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.13
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
textstat==0.7.3
thinc==8.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
tornado==6.2
traitlets==5.9.0
transformers==4.29.2
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.8.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.22.0
verspec==0.1.0
vine==5.1.0
wandb==0.16.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
weasel==0.3.4
websockets==11.0.3
wheel==0.38.4
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.0.4
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
fastapi-sqlalchemy==0.2.1
